+++ config/config.ini
@@ -1,8 +1,8 @@
 [app_setting]
-version = v2.6.4
+version = v2.7.4
 
 [model_providers]
-provider_list = openai, ollama, mistralai, localai, zhipuai
+provider_list = openai, ollama, mistralai, localai, zhipuai, anthropic
 selected_provider = ollama
 
 [openai_llm_models]
@@ -26,6 +26,10 @@ base_url = http://localhost:8080/v1
 [zhipuai_llm_models]
 model_list = glm-4
 selected_model = glm-4
+
+[anthropic_llm_models]
+model_list = claude-3-opus-20240229
+selected_model = claude-3-opus-20240229
 
 [embedding_model_providers]
 provider_list = huggingface, ollama
