+++ qa_model_apis.py
@@ -9,6 +9,7 @@ from dotenv import load_dotenv
 from langchain.callbacks.streaming_stdout import StreamingStdOutCallbackHandler
 from llama_index.llms.openai_like import OpenAILike
 from langchain_community.chat_models import ChatZhipuAI
+from langchain_anthropic import ChatAnthropic
 
 # read from the config.ini
 config_path = os.path.join('config', 'config.ini')
@@ -46,6 +47,11 @@ def get_chat_model(provider, model_name=''):
         return ChatZhipuAI(
             model=model_name,
             temperature=0.5,
+        )
+    elif provider == 'anthropic':
+        load_dotenv()
+        return ChatAnthropic(
+            model='claude-3-opus-20240229'
         )
     else:
         raise ValueError(f"Unsupported model provider: {provider}")
